move testWait as need to enable PILOT_ENABLE_CONFIG_DISTRIBUTION_TRACKING
